+++ src/main/java/com/example/ControleAlmoxarifado/dto/notaEntrada/CriacaoNotaEntradaRespostaDTO.java
@@ -2,6 +2,7 @@ package com.example.ControleAlmoxarifado.dto.notaEntrada;
 
 import com.example.ControleAlmoxarifado.model.Fornecedor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.HashMap;
 
@@ -9,6 +10,6 @@ public record CriacaoNotaEntradaRespostaDTO(
         Long id,
         Fornecedor fornecedor,
         LocalDate dataEntrada,
-        HashMap<String, Double> materiais
+        HashMap<String, BigDecimal> materiais
 ){
 }
