+++ src/main/java/com/example/ControleAlmoxarifado/dto/requisicao/CriacaoRequisicaoRequisicaoDTO.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.HashMap;
 
@@ -14,7 +15,7 @@ public record CriacaoRequisicaoRequisicaoDTO(
         String setor,
 
         @NotEmpty(message = "A lista de materiais não pode ser vazia!")
-        HashMap<Long, Double> materiais,
+        HashMap<Long, BigDecimal> materiais,
 
         String status
 ) {
